server:impl: verify FiveM tickets
This was a hard decision to make, especially as it means that any service downtime will result in authentication failures.

However, the recent spread of 'cracked' FiveM versions necessitates this, to keep the project out of the crosshairs of Rockstar/Take-Two Interactive, who might think that the 'cracked' copies will necessitate shutting down the original project.

In the future, this verification will be distributed across a few servers and domains - for now, this'll have to do, and hopefully it won't break for everyone.