launcher: fix autodetection for Steam copies
Steam GTA V copies don't use \Grand Theft Auto V\InstallFolder in the registry, but use \GTAV\InstallFolderSteam instead, directing to "%INSTALLDIR%\GTAV".

This value seems to not be used anywhere, as it uses a non-existent path, however it is still set by installscript.vdf, so we read this value instead.