server: manage cpr futures in a cleaner fashion in PERFORM_HTTP_REQUEST_INTERNAL
as the std::future destructor will cause a blocking wait on the future, this function was inherently blocking.

this patchset changes the lifetime of the std::future to only be destructed after the future completes.

we can not destruct the future within the callback, for this will cause a deadlock where the future waits for itself to have completed, so we track a list and clean it up on the next HTTP request.