http:client: use a queue for curl_multi_add_handle
The existing mutex was causing PatchStreamingPreparation to hitch when the game called EnsureFetched on a NotFetched streaming handle while cURL was processing another request.

Using a queue, while not kicking the thread directly (and therefore adding extra latency to requests), does get rid of hitching on the main thread when adding requests, which results in smoother loading of multiple streamed resources at once.