Update building.md
Added additional section for setting up testing enviroment. All those steps are scattered in different places (Discord, forum) and make it tough to properly build FiveM client for someone new to all of this. I've added all the things that should be sufficient to test modified client. Hopefully, this will encourage more people to try building the project and contribute (it was a big discouragement for me having to look for those additional steps not knowing they even existed in the first place)