fix: move non-parroting ack behavior to client
Since the entity was being reassigned by server logic with an *equivalent* check to the one added earlier, this check would fail and the fix logic would not apply.

This fixes that issue by making the *client* responsible for not parroting the server-induced entity deletion, thereby allowing the earlier fix to be triggered as well.