fix: server: treat uniqifier changes as new entity
If a client removes and recreates an entity fast enough, both last and new delta states might contain the exact same object ID, but referring to a different entity. Evidently, this'll result in the client rejecting state updates with a wrong uniqifier, and *worse*, the client rejecting removals as well.

Instead, we'll make a uniqifier change count as both a removal and a create - and change shouldSteal handling as well to not steal an entity which is actually different.