fix(onesync): object id reuse and redeletion fixes
Deletions didn't have a uniqifier, which is stupid, as uniqifiers existed for latency and reuse and all.

Also, make an object ID not reusable when the server still needs to tell us it got deleted.

Initial testing showed a marked reduction in deleted peds (1 player, 340ms latency), and no entity left behind when leaving scope and population.