fix(onesync/client): don't sync netobjs that aren't in the scene
The game often will create temporary entities, which do get registered with the object manager, but are not meant to be shown, interacted with or replicated until a later time, or get deleted instantly.

We also defuse the earlier delay hack as this should be fixing the root cause instead.