feature(streaming/five): cancel doomed streaming requests
If after many retries a streaming request returns as failed from RCD, we'd normally fail the game entirely.

Given the prevalence of connectivity issues, user error, corruption, user-induced corruption and other such weird things, instead we'll try to cancel the initiated request and it's dependent tree gracefully and safely.

This might lead to scripts that indiscriminately wait for model load hanging - it's still TBD whether errors should be propagated to script or if a user-visible warning should be shown.