tweak(devtools): remove frame time warning, for server owners apparently want 'their' players to suffer in low FPS
If nobody is going to optimize their servers as a result of this, and still continues whining even 4 days later, this gentle nudge doesn't help at all. It's a shame that end users are unable to ever provide info, server owners are always shifting every single blame to us, but yet aren't helpful at all and oftentimes outright hostile instead.