feat(clr): add ClientScript and ServerScript, allow missing types
This change adds ClientScript and ServerScript classes that are present
in the respective environments to differentiate client and server sided
scripts. This also allows the C# and F# compilers to unambiguously link
against the correct libraries when referencing both assemblies.

In addition, we handle an exception that is thrown when loading types
with missing assemblies, e.g. ServerScript is missing on the client.

See also https://forum.cfx.re/t/2006775