fix(net): prevent enet connection without connection token
We use the `data` field in enet_host_connect to pass a derivative of the connection token, which then is checked *early* in ENet connectivity (before allocating a temporary peer).

This should mitigate the scenario where fake peers either spam UDP packets to not get acknowledged, or acknowledge without ever sending an initial connection.