tweak(conhost): FPS counter rework
While running with Steam, it became apparent that the FPS counter in the
Steam overlay only updates once a second, instead of the 'every frame'
rolling average used in drawFPS/drawPerf.

Since users probably expect this counter to not flicker wildly, we're
moving it to the same once-per-second update cycle.

Code for cl_drawFPS and the FPS metric in cl_drawPerf was duplicated,
and this change set also unifies that.