fix(net/five): wrong pattern since 247123f817a6b81abd32bbc34fcc8806b41e1f75
As @GottfriedLeibniz mentioned, this pattern points into some replay
editor export code in all supported builds.

The original pattern appears to be from build 323, and I guess it was
only checked for matches in newer builds, but not if the matches were
pointing to the correct routine.

The original code in 'LookAlive' that was patched seems to be Arxan-
mutated in b1604 and above at the least, but we can (probably) safely
return_function the inner call on b2060 and above since there's only two
callers to this function, both of which are the skipover from b323.

This commit is *not* including a 1604 check, which may be a mistake if
there is a code guard on the function itself.