fix(client/launcher): no /STACK for main .exe
The last commit (c0a64ecbecd9aea8b40ce7192d2854c182d274f9) was found to
also have the stack override for the main FiveM.exe in `release` builds.

This was apparently due to a confusing `filter {}` statement in the
project file.

Even though this won't 'prevent' an unneeded launcher bootstrap even on
beta/prod, we should correct this.