fix(net/five): broken settings restart in net since 8b12f416ea42f4ba52848f0d2db06ad49cdc68e1
The fix commit mentioned in the subject *incorrectly* assumed there were
two places to patch, and that return_function on this function would
correctly 'suppress' the network disconnect.

This is incorrect: the 2nd Arxan-obfuscated caller is actually a native,
and overwriting with return_function won't execute the `else if` used
when the game is trying to exit/restart due to a setting change.

The second patch from 247123f817a6b81abd32bbc34fcc8806b41e1f75 simply
does not exist anymore in newer builds, nor was it ever required (as the
patches were tested on b1604+, each of these referencing broken replay
editor code for the patch).