fix(ci): unfuck Git post-'CVE-2022-24765' 'fix'
Git broke compatibility in a 'embargoed' 'security' 'fix' for an invalid
'vulnerability' which would impact a total of zero users at best, but in
the process broke every CI configuration in the world, with no obvious
way to revert to the old version as it followed the usual 'embargoed'
security fix procedures including zero pre-announcement of breaking
changes and replacing old versions even in old branches.

As it's easier to just patch the binary of the new version (20 days ago
by the way, just only pushed and forced on everyone ~at the time of this
writing) than to figure out a build environment for Alpine packages for
3.13, and in the Git developers (including @dscho who seems to be the
author of this amazing commit) their amazing, infallible wisdom, there's
no option to enable the old configuration or have the new configuration
as an opt-in or *anything* that'd have made sense, we're just patching
the binary.

Refs git/git@bdc77d1d685be9c10b88abb281a42bc620548595.

I'd reference an issue, too, but there's none. No PR either.