fix(gamestate/five): dummy object stub ABI
This broke as of 0fa9425436b32308fecdfb7b3d50e1c13b9f3f64, as reported
only recently in #1452 as a *reproducible* case of 'blowing up a gas
pump leads to the pump instantly being recreated'.

It turns out that the initial fix made a few mistakes in the ASM stub:

    1. `al` was overwritten with the post-restore `al`, and
    2. `rax` was saved in the scratch space for the callee, so it would
       often contain nonsense, causing 'Player Wants Control' to always
       be true.

Fixes #1452 - at least for the reproducible regression case, hopefully
for all cases however.