feat(game/five): 'pure' mode to disable modified game files
This is an early prototype of this functionality, and it should not be
expected to be fully functional and free of bugs.

It's also merely a suggestion to the client, 'cheat' injections can and
will ignore this state, which eventually may get picked up by private
components/patches for compliance enforcement ('adhesive').