fix(device/five): PathFilteringDevice prefix mistake
This was reported as 'with sv_pureLevel set, I'd have "pool exceeded"',
and investigating this showed that it compared filenames to ":/data/..."
instead of the correct trimmed filename of "data/...".

This also fixes the *other* path filtering in non-pure mode for files
such as 'materials.dat', 'trains.xml', 'dlclist.xml' etc. that are known
to potentially lead to (remote) game crashes.