tweak(client/minidump): add crashometry to crash zips
'Early-exit trap' failures only show relevant data in the crashometry
data, which leads to the counterintuitive behavior that users have to
send the report ID, and not the crash .zip like with other crashes.

This came up in a conversation recently, and indeed, there's not really
anything against adding the crashometry data to the crash .zip.