Allow scripts to access Game pause menu page
This lets 0x593feae1f73392d4 return 5 when the Game page is navigated to. I feel like it could be useful for servers, which would like to add items such as Rules, Information about the server etc. Tested it and everything seems to work, but if you find anything, feel free to let me know.