tweak(data/client): add b2699 Pause Menu tabs
Going through the new pausemenu.xml, these are the only "interesting"
changes I have found.

This commit adds Legal and Credit tabs in the GAME header (You need to
toggle contexts for one of these to show, can't have both at the same
time.)

Also added the 'Display Text Chat' setting item, however, this would not
work in older builds and would just appear as blank, so It can only be
enabled if the 'b2699' context is enabled (`PauseMenuActivateContext
("b2699")`).