fix(nui): handle 'wrapped' D3D11 textures in more cases
The GPU info collector, in some cases, may try to initialize a 'temp'
ANGLE D3D11 device with a specific set of flags leading to a 'shared'
1x1 texture being created, that subsequently gets passed into CreateRTV,
CreateSRV and CopySubresourceRegion.

Since these do not handle a Texture2DWrap, this then might lead to a
crash (FIVEM-CLIENT-1604-H0PK) in d3d11.dll.