fix(streaming/five): CInteriorProxy cache pattern for b2545+
The 'save' and 'load' pattern pointed to the exact same function, and
not the 'load' function.

This would lead to a state where the map data file would not get loaded
early but the interior proxies defined in the cache file would also not
exist, so the game would crash when trying to instantiate the MLO inst.

Reported on https://forum.cfx.re/t/4940689/6.