fix(nui): high-DPI-aware custom cursor
On FiveM at least, we're running in 'native' DPI mode, and this means
the cursor bitmap will be displayed 1:1 to output pixels, making for a
small cursor on high-DPI screens.

As a (hacky) fix, we will scale the cursor on load. Ideally this would
also reload/reapply when the window's DPI changes but that's out of
scope here.