fix(nui/core): D3D init deadlock with in-process GPU
Introduced in 524ec8ca56a2bfbd8c88fe3e1a3460e6a881351e, this would hit
a deadlock if this code ran with the Flush mutex held (such as when the
NUI in-process GPU option is enabled) as shared_mutex does not directly
have recursive ownership support.

Thanks to PichotM for the report + memory dump!