tweak(ext/cfx-ui): hide reviews that have a hidden first post
As reported in https://forum.cfx.re/t/5003108, sometimes handling a flag
in Discourse only marks the first post as hidden, which does not update
the post stream.

Instead, once a post loads, and it is marked as hidden, we just hide the
entire review.

This is not an ideal solution as it leads to brief flickering, but at
least will not show a 'removed' review for too long.