fix(build): python 'externally managed' error on linux
Apparently this behavior changed (pypa/pip#11663), and as seen in a
comment on pypa/pip#11381 this is *intended* to annoy people and 'break
workflows'.

Pesky open-source maintainers. Sigh.

The 'motivation' here is fun, too:

> This may pose a critical problem for the integrity of distros, which
> often have package-management tools that are themselves written in
> Python.

Who'd have thought that the 'everything is installed in a global place'
issue Linux distros have can't 'just' be solved by package managers as
suddenly the package managers can break themselves because, like, who
manages the package manager?

Hopefully the `--user` flag will work here and won't somehow not apply
as something in the build process uses a second user or something.