tweak(net/rdr3): workaround for task scenario crash
The task scenario getter crash happens when the game doesn't expect a function that meant to return `CScenarioInfo` from `rage::aiTask` for an unknown reason returns nullptr. It's almost impossible to reproduce on a local server, so we're trying to workaround this crash and output some maybe-useful notes when the potential crashing-case happens. For now we're only patching specific places where the crash appear to happen kinda often. Unfortunately, we can't be sure about no potential regressions introduced by this patch.