tweak(client): set GPU process as priority-boosted
In Windows releases above build 25xxx (targeted for release in 2024),
this fixes the underlying issue with GPU priority that usually led to
the 'in-process GPU' flag being needed.

In Windows releases before 25xxx (11 21H2, 11 22H2 and potential 2023
releases of 11) this is also implemented but it only affects the CPU
priority as the code path taken doesn't use the dxgkrnl priority boost
callback.

This technically is a LAF (limited-access feature) but we can't request
access to it the 'official' way since the email alias outright doesn't
work outside of Microsoft so instead we override the LAF directly in the
PEB.

See https://forum.cfx.re/t/4977782/7 for more info.