fix(glue): disable packfile signing until the new PKI system exists
It appears the 'NVE' graphics mod, which was in the pilot program for an
older signing system, has bundled a set of weapon meta changes as an
'optional add-on' under the name 'r_NVE_Weapons_Overhaul_Pack.rpf'.

This technically counts as a cheat, breaking the point of pure level 1.

As a hotfix, we're removing the signing key until the new infrastructure
for signing, tracking modified files (in case some files that impact
gameplay get signed inadvertently), and scoped revocation (so in this
case, only this add-on package would get blocked, and only for pure mode
- not the entire game) is finished.